% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRC_fitting.R
\name{fit_AC50}
\alias{fit_AC50}
\title{Fit AC50s to dose response data}
\usage{
fit_AC50(df, conc_um, readout_value_pctAC, best_fit = TRUE,
  abs_acx_level = 50, a_diff_thresh = 30, rsq_threshold = 0.5,
  constr_lower = c(-50, -50, log10(1e-06), 0.1), constr_upper = c(500,
  500, 3, 10))
}
\arguments{
\item{df}{dataframe containing concentration and readout columns}

\item{conc_um}{treatment concentration in uM}

\item{readout_value_pctAC}{readout value as percentage of an active control}

\item{best_fit}{by default, only the best fit is reported. If set FALSE, all three fits are reported (the best fit is reported with fit_sel=1 )}

\item{abs_acx_level}{threshold for absolute AC50s, default is 50 percent activity}

\item{a_diff_thresh}{if amax-amin < a_diff_thresh, fit is considered constant}

\item{rsq_threshold}{r squared threshold for parametric and non-parametric fits to be considered}

\item{constr_lower}{list of lower limit constraits for parametric fit for a0, ainf, ac50, hill, default c(-50, -50, log10(1e-06), 0.1). Use -Inf for no constraints.}

\item{constr_upper}{list of upper limit constraits for parametric fit for a0, ainf, ac50, hill, default c( 500,  500,  3,  10). Use Inf for no constraints.}
}
\value{
data.frame with AC50s and fitted parameters
}
\description{
Function to calculate best fits for dose response data. Three fits are attempted, constant, 4-parameter logistic regression, 
and a nonparametric fit. The best fit is selected based on a heuristic.
}
\examples{
see vignettes


}
