% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lincs_GR50s}
\alias{lincs_GR50s}
\title{GR50 growth inhibition data for EGFR inhibitors in MCF10A cells}
\format{A data frame with 24192 rows and 13 variables:
\describe{
  \item{GR50.cellLine}{cell line for which data was measured}
  \item{GR50.smallMolecule}{compound}
  \item{GR50}{GR50 as defined in the Hafner paper}
  \item{GRmax}{GRmax as defined in the Hafner paper}
  \item{GR_AOC}{GR_AOC as defined in the Hafner paper}
  \item{GEC50}{GEC50 as defined in the Hafner paper}
  \item{GRinf}{GRinf as defined in the Hafner paper}
  \item{GR50.Hill}{GR50.Hill as defined in the Hafner paper}
  \item{GR50.r2}{GR50.r2 as defined in the Hafner paper}
  \item{GR50.pval}{GR50.pval as defined in the Hafner paper}
  \item{cpd_id}{standardized compound name}
  \item{cl_id}{standardized cell line name}
  \item{GR50.e}{estimated GR50s. For cpds with flat curves, more extreme GR50 potencies were estimated than just the maximal concentrations.}
}}
\source{
\url{http://lincs.hms.harvard.edu/db/datasets/20252/results}
}
\usage{
lincs_GR50s
}
\description{
A dataset containing GR50s activities of EGFR inhibitors in MCF10A cells 
The data and algorithm is described in more details in Hafner et al 2016 (doi: 10.1038/nmeth.3853)
}
\keyword{datasets}
