% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lincs_tpx2_sig}
\alias{lincs_tpx2_sig}
\title{Dose-dependent changes in gene expression of an TPX2 signature in MCF10A cells upon treatment with EGFR inhibitors}
\format{A data frame with 24192 rows and 13 variables:
\describe{
  \item{well_type}{classify wells into one of "SA" for sample treatment, "AC" for active control, and "NC" for neutral control}
  \item{gene_signature}{annotation of gene signature for comparison of datasets}
  \item{cpd_id}{standardized compound name}
  \item{cl_id}{standardized cell line name}
  \item{id}{original id from L1000}
  \item{SM_Dose}{compound concentration in uM}
  \item{CL_Center_Specific_ID}{original cell lines id}
  \item{SM_Name}{name of compounds}
  \item{SM_Pert_Type}{pertubation type as annotated in L1000}
  \item{SM_Time}{incubation time}
  \item{det_plate}{assay plate}
  \item{det_well}{plate well id}
  \item{...}{many columns named after genes containing changes in normalized gene expression}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE70138}
}
\usage{
lincs_tpx2_sig
}
\description{
A dataset containing the concentration dependent changes in expression 
of the genes of an TPX2 signature upon treatment with EGFR inhibitors at 
6 different concentrations. The data is a subset of the L1000 data set
More details on the full dataset in Subramanian et al., Cell, 2017 (10.1016/j.cell.2017.10.049)
}
\keyword{datasets}
