% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpda_ac50s.summary_readouts.R
\name{signature_scores}
\alias{signature_scores}
\title{Signature scores}
\usage{
signature_scores(df, readouts, keep_cols, well_type, mv_methods = c(),
  threshold = 3, keep_n_most_responsive_readouts = 0,
  keep_readouts = c())
}
\arguments{
\item{df}{wide dataframe with one row per treatment and multivariate readouts in columns}

\item{readouts}{list of columns containing multivariate readouts to be summarize, e.g. c("col1", "col2", ...)}

\item{keep_cols}{list of columns with metadata to report in results, e.g. c("col1", "col2", ...)}

\item{well_type}{column specifying samples as "SA", neutral controls as "NC", and active controls as "AC"}

\item{mv_methods}{list of multivariate summary methods to compute. Options are:
\itemize{
 \item "maha_AC" mahalanobis distance to the AC,
 \item "maha_NC" mahalanobis distance to the NC,
 \item "euc_AC" euclidean distance to the AC,r
 \item "euc_NC" euclidean distance to the NC,r
 \item "cor_p_AC" Pearson correlation to the AC,
 \item "cor_s_AC" Spearman correlation to the AC,
 \item "dot_p_AC" dot product to the AC,
 \item "cos_AC" cosine to the AC,
 \item "vec_norm" vector norm,
 \item "scalar_projection_AC" scalar projection to the AC,
 \item "cos_weight_AC" cosine to the AC multiplied by max(1, mean(abs(readout_values)) / threshold)
 \item "num_readouts_changed" number of readouts with abs(values) above threshold.
 }
 Default are all methods except the mahalanobis distances.}

\item{threshold}{threshold used in cos_weight_AC and num_readouts_changed. Default is 3, assuming the readout values are normalized as zscores or rscores, so that a threshold of 3 means 3 standard deviations.}

\item{keep_n_most_responsive_readouts}{number of the original readouts to keep which are most responsive to the AC condition}

\item{keep_readouts}{list of columns of the original readouts to keep for results, e.g. c("col1", "col2", ...)}
}
\value{
Dataframe with keep_cols, potential original readouts, and multivariate readouts with
readout_value the raw value
readout_value_pctAC the readout value normalized as percentage of the AC
readout_value_pctACNC the readout value normalized to NC and as percentage of the AC
}
\description{
signature_scores calculates a set of summary methods for the effect of a treatment on multivariate readouts. 
Summary methode quantify effects based on the signature data alone, or relative to a neutral or active control signature.
}
\examples{
see vignettes

}
