% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thp1_cAMP}
\alias{thp1_cAMP}
\title{Dose-dependent changes in cAMP levels for THP1 cells treated with beta agonists}
\format{A data frame with 24192 rows and 13 variables:
\describe{
  \item{row}{row on plate}
  \item{col}{column on plate}
  \item{assay_plate}{assay plate}
  \item{well_type}{classify wells into one of "SA" for sample treatment, "AC" for active control, and "NC" for neutral control}
  \item{conc_um}{compound concentration in uM}
  \item{inchi_key}{inchi_key for compounds}
  \item{smiles}{smiles encoding of compound structure}
  \item{biol_replicate}{biological replicates}
  \item{readout_name}{name of readout}
  \item{readout_value}{readout value from assay}
  \item{readout_value_pctAC}{activity normalized to percent of the active control}
  \item{readout_type}{type of readout}
  \item{cpd_name}{name of compound}
  \item{readout_set}{annotation for comparison with gene signature results}
}}
\source{
NIBR
}
\usage{
thp1_cAMP
}
\description{
A dataset containing the concentration dependent changes in cAMP levels
of THP1 cells treated with compounds.
}
\keyword{datasets}
