% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thp1_gene_rscores}
\alias{thp1_gene_rscores}
\title{Dose-dependent changes in gene expression for THP1 cells treated with beta agonists}
\format{A data frame with 24192 rows and 13 variables:
\describe{
  \item{row}{row on plate}
  \item{col}{column on plate}
  \item{assay_plate}{assay plate}
  \item{well_type}{classify wells into one of "SA" for sample treatment, "AC" for active control, and "NC" for neutral control}
  \item{conc_um}{compound concentration in uM}
  \item{inchi_key}{inchi_key for compounds}
  \item{smiles}{smiles encoding of compound structure}
  \item{biol_replicate}{biological replicates}
  \item{gene}{gene measured}
  \item{rscore}{rscore: number of robust standard deviations of gene expression away from neutral controls}
  \item{rscore_HKnorm}{rscore normalized to housekeeing gene }
  \item{cpd_name}{name of compound}
  \item{readout_set}{the 14 genes were measured in two sets of 7 genes + housekeeper}
}}
\source{
NIBR
}
\usage{
thp1_gene_rscores
}
\description{
A dataset containing the concentration dependent changes in expression 
of 14 genes and a housekeeper gene upon treatment with compounds at 
8 different concentrations
}
\keyword{datasets}
